# Conflicts:
#	data-manager/src/main/java/com/data/controller/GameController.java
#	data-manager/src/main/java/com/data/mapper/GameMapper.java
#	data-manager/src/main/java/com/data/service/GameService.java
#	data-manager/src/main/java/com/data/service/impl/GameServiceImpl.java
#	data-manager/src/main/resources/mapper/GameMapper.xml
#	manager-admin/src/main/java/com/manager/controller/data/GameController.java
#	manager-admin/src/main/java/com/manager/openFegin/DataService.java

+++ data-manager/src/main/java/com/data/controller/GameController.java
@@ -28,7 +28,7 @@ public class GameController extends BaseController {
      * 获取游戏下拉选项
      */
     @PostMapping("/option")
-    public AjaxResult list(Long tid) {
+    public AjaxResult list(Integer tid) {
         return AjaxResult.success("查询成功", gameService.getGames(tid));
     }
